Added docbook based manual page for libcouchbase_create.
This also adds the basics for the new documentation approach to
the project.

This does not immediately remove all of the other man pages, but
rather adds the new docbook based approach.  With this change,
man pages will be added here in docbook and reused in other
projects.

By shifting to this style of documenting the API, it can be reused
in other documentation artifacts and can be easily crossreferenced
to other client APIs.

Change-Id: I559d4ad28c3e45c1a91ab4f33b6de0354ebb2645
Reviewed-on: http://review.couchbase.org/11353
Tested-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>