Add cbc send and receive
cbc send converts a tap dump to a stream of set and delete messages
written to stdout. cbc receive should convert this stream so that you
could apply it to the server. For now it discards the CAS field and
vbucket identifier, so that the item if redistributed into the
vbuckets on the other cluster.

Conflicts:

	Makefile.am
	tools/cbc.cc

Change-Id: Ie8dfc779fd731f2f3e8155909d7e6cce2045e5b1
Reviewed-on: http://review.couchbase.org/12216
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Matt Ingenthron <matt@couchbase.com>