Updated version and package metadata for 1.0.1.
The actual 1.0.1 in the package file happens based on the git tag
but we set it in package metadata manually and control the version
in configure.ac.

Note this release introduces divergence in the so-name and the
package version.  We've decided this is the appropriate thing to do
for this project.  As you may know, various approaches are taken by
various projects-- our rules are very strict on so-name and more
loose on product release version.

In the future, the first commit after a release should increment
this, but we didn't do that this time.

Change-Id: I5944ca486ac80a3dd92e68823c01fe19f697b471
Reviewed-on: http://review.couchbase.org/13214
Reviewed-by: Jan Lehnardt <jan@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>