Bundle the protocol definitions for memcached to make it easier to build
The protocol headers is used to get the struct definitions and the
command id's to make it easier to build on most platforms. Those files
doesn't change that often, and it doesn't really make sense to require
the user to build and install the memcached server just to get the
headerfiles to build the client ;-)

I've added the COPYRIGHT file from the memcached project in the same
directory as the headers.

Change-Id: I73a469704accd333cef7ead7a4a40701a704bd1c
Reviewed-on: http://review.couchbase.org/13633
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>