[CCBC-65] Memory leak in libcouchbase_create / libcouchbase_destroy
When the client calls libcouchbase_create and then later calls
libcouchbase_destroy, some memory is leaked. For a client with many threads
that open and close couchbase sessions, this will lead to memory exhaustion.
The leak is caused by the use of strdup() in setup_boostrap_hosts(). These
duplicated strings are not free'd in libcouchbase_destroy. The leak can be
seen by using valgrind.

Thanks to James O'Connor

Change-Id: I376d0c767ddb9335feae6d7c28cd2e391209101c
Reviewed-on: http://review.couchbase.org/16032
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>