CCBC-115 Return zero from do_read_data() if operations_per_call reached
The operations_per_call limit was introduced to prevent from freezing event
loop. But in the function variable rv could store two different results and
in case of reaching this limit it is returning number of the processed
records, which is wrong. The function should return either zero (success) or
non-zero (failure).

Change-Id: Ia576739de35eccaa1b7e8c241ba43ad3e29d57a1
Reviewed-on: http://review.couchbase.org/22310
Reviewed-by: Mordechai Nunberg <mnunberg@haskalah.org>
Reviewed-by: Michael Leib <michael_i_leib@yahoo.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>