CCBC-166: Fix manpage generation
We shouldn't assume that everyone who builds libcouchbase
has a working setup of a2x (and it's painfully slow on macosx
if installed through homebrew). A better solution would be to
generate the files on the developers machine and commit the
generated manpage.

Running "make regenerate-mapage" rebuilds all manual pages
and create a BSD version and a SVR4 version of the man page
(they use different section numbering). Given that BSD don't
structure the manpages into more sections than a number, write
your manpage in the "svr4" layout and it'll be moved to the
correct subdirectory during "make regenerate-manpage"

Change-Id: Id5d67e08527cdc242480328e96a70f9377815eab
Reviewed-on: http://review.couchbase.org/23749
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>