Refactor boilerplate I/O for non-critical paths
This patch adds an 'easy' mode - so I/O consumers which aren't in the
critical path (i.e. config updates and HTTP handlers) can use a single
group of functions for both v0 and v1 I/O.

Note that the paired v0/v1 I/O is still fully supported and is used by
the 'server' code. We might want to switch to it if it doesn't impact
performance too much :)

Change-Id: Ie199e184bd25c4b14f4a13540ef5c9ce85e1b003
Reviewed-on: http://review.couchbase.org/31424
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>