Provide 'last_error' connection field
This is mainly useful during the connect sequence and can be helpful in
debugging various errors along the way. The LCB error at this point is
rather unhelpful (something we might want to change in the future) and
does not intend to expose the full array of OS errors anyway.

Change-Id: If537289fa2ead15491ee07f3fea9aee8b7a21790
Reviewed-on: http://review.couchbase.org/28133
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>