Reset the server's buffers upon reconnection
This fixes a crash experienced when requesting a new read with the
previous buffer still in tact.

This was exposed by calling lcb_failout_server on a timeout error while
maintaining the same server struct. This path is only followed for
failed negotiations but is the only time in code where we re-create the
socket.

Change-Id: Iedaed57ddf689189cc597eea29165d861ed14c97
Reviewed-on: http://review.couchbase.org/29681
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>