Refactor SASL negotiation
NOTE: the sockrw_flushed check has been removed from
lcb_server_has_pending because it causes test failures and is not needed
in this function. Specifically there is never a case when the server
has data to flush and does _not_ have a pending command somewhere. Even
if it did have data to flush, we shouldn't hold up the event loop for
that anyway.

Additionally, the negotiation callback will reset the buffers, ensuring
that if the connection's output buffer is still pending, it will be set
on V1 plugins

Change-Id: I4f291aa819360235eaaa8fad7008772abceef496
Reviewed-on: http://review.couchbase.org/31406
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>