Add CCCP Support
CCCP Support is introduced with a new creation structure (v2) which
allows the specification of explicit memcached nodes.

Unfortunately the tests still need to adjust as it will try to connect
to the default port -- and the mock being none the wiser, isn't setting
the explicit 'mchosts' parameter. The workaround is to disable any local
Couchbase cluster before running the tests.

Change-Id: I5fff6e2346a4fde2ed7fe8191d3b40bd470671db
Reviewed-on: http://review.couchbase.org/31428
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>