Add more semantic error functions
This commit adds error 'classifiers' to existing error codes while
maintaining full API/ABI compatibility with existing behavior. The idea
of this commit is to migrate from checking indidivual error codes (i.e.
LCB_ETIMEDOUT or whatever) towards checking error classifiers (i.e.
LCB_EIFNET(e) which function similarly to the WIF* macros for the
POSIX wait() family of functions.

It would be nice to have a file which could maintain backwards
compatibility for older versions of the library (i.e. one that can be
included into the source code of older SDKs) so that we don't have to
rely on the 'lcb_get_errtype' and 2.3

Change-Id: I57e5f595fa0e08f52a64388ab3f76f51aa3deaad
Reviewed-on: http://review.couchbase.org/33406
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>