Map NOT_MY_VBUCKET calls to LCB_ETIMEDOUT
While this is technically incorrect, the only time we allow a
NOT_MY_VBUCKET to reach the response handler code here is during the
event of a timeout where the command is no longer retried. This
simplifies application development in allowing this failure to be
treated as a normal "Timeout" - with all its actionable dependencies.

Change-Id: Iae9806eb0a831d492f06c0fdd175ccdf312c28e9
Reviewed-on: http://review.couchbase.org/34650
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>