m4/couchbase.m4: always disable aliasing for gcc
Strict aliasing is a dangerous option which GCC unfortunately has
enabled by default in c99 mode (which is what we use when employing
gcc). Many larger projects already disable this.

In the future we might also want to embed these options as pragmas for
GCC within the relevant affected files so we are not at mercy of the
compilation options

Change-Id: I381d1ff308ef564f6cd122bc3890424799c320c1
Reviewed-on: http://review.couchbase.org/36227
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>