CCBC-383: Provide bootstrap success/failure API
This commit adds an API to determine if bootstrap was successful or not.
A callback is provided for asynchronous use cases which is invoked
exactly once with either LCB_SUCCESS or the error code.

Another API is provided to be more friendly to synchronous,
lcb_wait()-based use cases where one may call lcb_get_bootstrap_status()
after the initial lcb_connect/lcb_wait sequence to determine if
bootstrapping succeeded.

Change-Id: Ic322117a639c6474b6602d764026ab89378d78df
Reviewed-on: http://review.couchbase.org/36229
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>