iotests: provider wrappers around lcb_create()
Because we now use the terse URI which the Java mock does not handle too
nicely, we are forced to disable the terse URI or risk random oddities
taking place in tests.

Provided is a doLcbCreate() function which wraps creation with various
'ctl' settings needed to ensure proper functionality within the test
environment.

Change-Id: Ibdd3eb16a1561056afcd0ee8d4dad08cbc655c05
Reviewed-on: http://review.couchbase.org/37615
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>