Sync up autotool configuration
This adds support for the newer files which were added within the
library recently and also introduces 'snappy' support. Since autotools
behaves oddly with linking in C++ archives I've also added a
'config/dummy.cc' file which is added to the main library's source list
if snappy is being built. This makes it clear to libtool that we are
building a C++ library.

Change-Id: I1a947305401231fd395fdbd243fa5b7b91e9b805
Reviewed-on: http://review.couchbase.org/38296
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>