Pack and reorganize settings structure
This commit makes the settings structure smaller by about 50 bytes by
reducing some field's sizes and converting boolean or small value fields
into bitfields. Since the settings structure is access by almost
everything within the library and is within its hot path, we want it to
be as efficient as possible to access - and this means ordering commonly
used fields in the beginning and making the overall structure smaller.

Change-Id: Ib672d84367d6fbad82dc495ab6f84a57c72d727d
Reviewed-on: http://review.couchbase.org/38299
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>