Add API for determining which features are compiled into the library
This is mainly so folks can do runtime checks. Currently only checking
for SSL or snappy is supported.

This commit also moves the APIs (including the lcb_get_version() API)
back into the main header file, so that one can use a 'static'
configuration.h. In the future we might want to completely enable a
'configless' compile.

These changes are also reflected in the 'cbc version' output

Change-Id: If2bf46681f131cc2ab0a5b38cfd968c5c4e962ab
Reviewed-on: http://review.couchbase.org/38523
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>