Make cntl.h smaller
This file contains most of our settings and is the point of entry to
seeing how the library may be configured (barring API documentation). In
order to make sure users have as simple a time as possible browsing this
header we should try to make it smaller without actually removing any
valuable information. These changes consist of:

- Moving deprecated codes to deprecated.h
- Making the argument description table less wordy
- Make enum/structure declarations more compat

Change-Id: Ibcb82aa6b7860317734251be915040144816d6ac
Reviewed-on: http://review.couchbase.org/38809
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>