CCBC-226: Use keepalive for views
This changeset creates a socket pool (lcbio_MGR) instance for HTTP
connections. In order to avoid situations where a pooled connection is
no longer functional, its use scope is limited to view requests and the
socket is only placed back into the pool if the underlying parser object
(via http_parser) indicates that the connection is keepalive-enabled.

Change-Id: I063ffcc67feee990d489816dbbd184e9affe7ad3
Reviewed-on: http://review.couchbase.org/39146
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>