Deprecate lcb_timer* and replace with lcbio_timer
lcb_timer has been deprecated before, but since we had a lot of internal
usages of it, we didn't actually mark it as deprecated to the
preprocessor. This commit replaces all usage of lcb_timer within the
internal code with that of lcbio_timer. The lcb_timer code itself still
remains but is now sectioned off in a single file (timer.c) and isn't
touched by anything within the library.

Change-Id: Ibfc05e2a0f2ee2ff8a5461988b426e56152e2666
Reviewed-on: http://review.couchbase.org/39215
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>