Always include configuration.h in distribution
The script would previously include only the files which were found in
the 'include' directory. If re-running gensrclist.pl after a successful
build (which is what I often do in my environment) the problem did not
appear as configuration.h already existed. However running from a clean
tree the configuration.h file is never picked up.

The solution here, is to mention it manually :)

Change-Id: If1b99e76c5c5ffc102a83f5e2d864dd5c882da8e
Reviewed-on: http://review.couchbase.org/39257
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>