Doxygen: Don't generate manpages
Unfortunately the current manpage output is lacking - we might re-enable
this in the future if nicer output can be generated. Note that this can
always be reenabled by the user by doing:

(cat Doxyfile; echo GENERATE_MAN=yes) | doxygen

Note that this does not effect the manpages for the command line tools,
only the programming APIs which doxygen generates by default.

Change-Id: I07e17a922ea8162fa4f63290953aa1e43b69b217
Reviewed-on: http://review.couchbase.org/39316
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>