CCBC-475: Add SSL error codes and messages
This commit introduces a bit more correctness when returning SSL-related
errors to users.

Two new error codes have been introduced:
- LCB_SSL_ERROR: Returned when initialization of SSL fails
- LCB_SSL_CANTVERFIY: Returned when the server's certificate cannot be
  verified.

LCB_PROTOCOL_ERROR is now returned when it seems we are not talking to
an SSL host.

Unfortunately openssl's error codes are not very well documented so it
is still likely for other cases to slip through the cracks in this
commit.

Change-Id: Ie85ba2a555224cc43cdd27764e79cfada9c79ecc
Reviewed-on: http://review.couchbase.org/39320
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>