autotools: disable distcleancheck
distcheck is a very helpful way to verify the distribution files are in
tact and that the package is configured correctly. However by design we
leave some files in the source tree (specifically, ones we've
downloaded) to avoid re-downloading them in the event of make distclean
&& ./config/autorun.sh && ./configure - which is a good way to clean and
rebuild the distribution.

This file disables the distcleancheck rule which would otherwise return
an error code because those downloaded files are present.

See
http://www.gnu.org/software/automake/manual/html_node/Checking-the-Distribution.html
for more information

Change-Id: Ieaca5d4ae0d2c82ea1a9a1c342019747e6c8d7cb
Reviewed-on: http://review.couchbase.org/39552
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>