vbucket: provide vbalternate() function
This function has a purpose similar to the older
vbucket_found_incorrect_master, except that it does not change the
vbucket state. Rather it simply tries to return a different index
than the default master.

As part of this change, the lcbvb_genconfig has also been modified to
generate unique indices for the master and replica slots for each
vbucket to accomodate the tests here (and function like a real config).

Change-Id: I44353375073735037dba54d55477d95316f0fc75
Reviewed-on: http://review.couchbase.org/39824
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>