Rework retry setting options
This changeset makes the retry setting be an integer (lcb_U32) rather
than a structure, making it easier to be assigned using the traditional
'ctl' functions in other SDKs.

Additionally, a new string alias ("retry_policy") has been added so that
the setting can also be modified using a string.

To accomodate these changes, a more generic string-to-integer mapping
subsystem has been put into place for cntl.c as well, so as to reduce
boilerplate.

Change-Id: I854ebf274f2b941c0b03a028ba6aa1671c9085e1
Reviewed-on: http://review.couchbase.org/39924
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>