retryq: never fail missing node items during config refresh
If a configuration is being fetched currently then there is a good
chance that the item will end up with a node to service it. This allows
a sane behavior of not keeping the command 'aimlessly' in the queue if
there is no configuration to be retrieved, while on the other hand
allowing the item to wait until the configuration is finally fetched.

To accomodate this change, a new bootstrap internal API was added that
allows to request a new config purely based on the last-refresh interval
rather than also taking the 'error count' into consideration. We'd like
to refactor the internal bootstrap APIs as well but now is not the time
:)

In light of these changes, this fix also modifies the default behavior
to "Fail Fast" on items missing nodes, since the item will wait
appropriately for a new config if one is indeed being requested.

Change-Id: I0cdae86aae45178203917a1a9ec5561b142e239e
Reviewed-on: http://review.couchbase.org/39956
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>