CCBC-484: Patch ruby IOPS for creation
This patch resolves an issue which would otherwise break the ruby client
implementation. This is because the SDK assumes niov will be 2 for all
I/O functions, but this is no longer true.

This patch surgically replaces the default send/recv I/O related
functions if it's detected (heuristically) that Ruby is being used.

Change-Id: Ieef83b4b7b92d9728c614c1e6d725d41183c734b
Reviewed-on: http://review.couchbase.org/40030
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>