Provide lcb_dump() to output internal state
Currently this function is considered volatile. It will dump the
relevant state information of the library to a `FILE*` pointer passed to
it.

This patch touches some other areas of the code, mainly so that existing
dump functions accept a `FILE*`, and that the relevant information is
output.

Another (undocumented) argument (--dump) has been added to cbc, mainly
to test lcb_dump()'s rudimentary functionality.

Change-Id: If58fc74088feb017ce6c776f8edf7accee9779aa
Reviewed-on: http://review.couchbase.org/40070
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>