Improve initialization documentation
The documentation for the connection string and such should be placed in
the beginning of the section, rather than as specific to lcb_create_st3.
This makes the details easier to find.

This commit also adds some other minor improvements, such as additional
documentation to lcb_set_cookie, and reorganizing of some of the
functions in the couchbase.h header file

Change-Id: I50f69913605cd326eeb470386bbcf2b677ef13f4
Reviewed-on: http://review.couchbase.org/40921
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>