cbc: Fix ignoring of cbcrc options
Some cbcrc options were previously being ignored due to their
corresponding Option structures not being marked as 'passed' or 'found',
but merely containing a default value. The code which would then apply
the options ended up ignoring these values because the passed flag was
off. This ensures that the pass flag is set when the defaults are loaded
from a config file.

Change-Id: I4705690e3a53eb3876790e0b932ccccbce09f1b0
Reviewed-on: http://review.couchbase.org/41300
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>