CCBC-519: rdb: wipe available blocks when done
This helps avoid a rather annoying memory 'leak' (when in fact we're not
really leaking, just allocating and keeping around way more than we
should be).

Note that I've still observed some smaller spikes in memory usage with
that same kind of workload, thus we might have more laying around.

Change-Id: Ice1234a354b9a395788a00a67441016eed68ed9c
Reviewed-on: http://review.couchbase.org/41842
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>