Only send out request body on PUT/POST
While technically the HTTP spec allows a request body with requests
other than `GET` (or more specifically, it does not explicitly disallow
it), it still causes confusion in ns_nserver with rather odd to debug
behavior.

Rather than propagate this behavior up the stack, let's ignore it and
try to do the right thing.

Change-Id: Ifaeb5c66bd9bd9826bb56f211387a177dc27325b
Reviewed-on: http://review.couchbase.org/42170
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>