CCBC-523: Disable support for SSLv3
This disables the usage of SSLv3 and by extension disables being
affected by the 'POODLE' attack.

This patch also will log to the screen with information about the
current SSL version and cipher being used.

Note that while the name of this method is SSLv23_client_method,
this actually allows support for all versions of the SSL protocol. We
explicitly disable support for SSLv2 and SSLv3 later on.

Change-Id: I08cb31a112b5ecf2c41f24c116fed0843c1a50b4
Reviewed-on: http://review.couchbase.org/42172
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>