Fix segfault when SSL is disabled
This would previously revert to using a macro which did not properly
initialize the error variable (leaving it as random garbage). Depending
on the contents of the uninitialized memory, this would either yield an
`LCB_SUCCESS` or something else that was odd.

This fix explicitly uses a different implementation of lcbio_ssl_new()
which ensures that a proper (new) error code is returned if SSL is not
built into the client library.

Change-Id: Ibab8729e2a75ccc4a42593c42f3828f994f2643f
Reviewed-on: http://review.couchbase.org/42241
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>