CCBC-331: Relocate commands on topology changes for memcached bucket
This simply ignores `hashkey` (considering as it's going the way of the
dodo) and simply rehashes the item based on their ketama values.

Interesting to note that since we maintain server connections in tact by
default, this new patch simply forwards commands to other servers
(despite them already being sent to the existing server), in order to
minimize the chance of receiving errors.

Change-Id: I171aaef70116cc166e4aa6956169b3dacf4f2d63
Reviewed-on: http://review.couchbase.org/42445
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>