Deprecate v1/v2 IOPS access internally
We just patched up a bug from older code which was using the 'v0'
fields. Let's avoid this in the future by disabling access to all the
'v1' and 'v2' specific plugins.

Note, non-internal code (i.e. LIBCOUCHBASE_INTERNAL is not defined) will
still work (and not get warnings), and the ABI structure will also
remain the same.

Change-Id: I42affa93f40f5db182bb8ebffc46f2758ad386e7
Reviewed-on: http://review.couchbase.org/43270
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>