ctx/ioC: Close older socket when the previous connect fails
This is a bug which was left undiscovered because existing ioC
implementations would not allow a connect() to be re-invoked on a socket
which had initially failed, so the effect remained the same. However
this seems not to be the case for all implementations, and would result
in the C_conncb being re-called in an infinite loop, since the next-call
to connect() would be invoked on the existing socket (which actually
does fail).

Change-Id: Ibed955ffcde6d7b55d9da181607761cb12935aa3
Reviewed-on: http://review.couchbase.org/43287
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>