Embed googletest into repository
This adds the google test source code into the contrib/ directory.

The googletest suite is _not_ available on external package managers by
design, the reason being that this library is meant to be embedded by
developers using it.

Since the google test distribution is rather large (1.3MB compressed), I've
stripped out some stuff (all build-related) that we won't ever used. As
a result, the compressed size is significantly smaller.

The most important impact of this change is that builders can now have
the tests for 'free', without having to download googletest during each
run.

Change-Id: I6b9f7800fedb3b2bd187681d2448bfbc1d39cb26
Reviewed-on: http://review.couchbase.org/43534
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Todd Greenstein <toddman@gmail.com>
Tested-by: Todd Greenstein <toddman@gmail.com>