connspec: don't fail if certpath passed without bucket
Despite the fact that the certificate path technically needs to be
percent-encoded, it's a common enough use case that we should try to
make it conforming. Each field itself can already be percent-encoded
(and thusly decoded by the library), so there is always the workaround
for folks who might actually have a '?' in their bucket or hostname.

Change-Id: Ib38cb2d775da08d06fe860ac2566aa367310c5ab
Reviewed-on: http://review.couchbase.org/44132
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>