CCBC-566: DNS SRV support
This adds DNS SRV lookups for bootstrapping. This is provided in two
forms:

1. When there is only a single hostname with no port. This is called
   implicit lookup
2. When the scheme `couchbase+dnssrv://` is used. This is an explicit
   mode and is considered internal to the library.

In both cases the input hostname will be selected and attempted for a DNS SRV
query. If it succeeds, then the hosts use for the query will be
returned. If the query fails, implicit mode will continue, using the
hostname as a couchbase node, whereas explicit mode will fail. Explicit
mode is useful for debugging the dns srv feauture itself.

Currently support depends on either Windows (which works "out of the
box") and libbind/libresolv, which come bundled with glibc on both OS X
and Linux. This may fail to function on very ancient systems; in which
event these routines are stubbed out with no-op equivalents.

Change-Id: I0d54b63c1e6cef64e5a09280343e9aba24ed7a80
Reviewed-on: http://review.couchbase.org/46177
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: buildbot <build@couchbase.com>