[OBSEQ] low-level lcb_observe_seqno3
This adds the protocol-level observe_seqno3. The higher level durability
commands will build on top of this command.

This commit depends on a new version of CouchbaseMock
(http://review.couchbase.org/#/c/46397/) and thus should not be merged
until the mock itself is updated - and we can use it for tests.

Tests for basic protocol handling for both the extra 16 bytes on
mutations as well as the OBSERVE_SEQNO command itself have been added.

Change-Id: I2ed5bbdad71889f85564dad5c423dfa79ad2bd7c
Reviewed-on: http://review.couchbase.org/46471
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>