CCBC-582: Timing fixes
This commit fixes timing behavior in various off-by-one measurement
intervals.

This fixes the issue of the bug in CCBC-582, accounting for 1-2k ms ops.
This also converts reporting of 1-10s intervals to be in the form of
millis mainly because of the range system and for finer accuracy (what
about 1.5s?).

Change-Id: Idf33bfd59c2df0e9b2c04a5105c2b973006500e0
Reviewed-on: http://review.couchbase.org/47947
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>